[DO NOT MERGE] JAVA 17 BWARE COMMIT
tsmm cleanup

more cleanup

llm cheaper

format DML Runtime Exception

remove logging of java version in bin/systemds

inc sliceline

cleanups

comment out scuro

Common Thread Pool

thread pool

vector api for worker

Federated Worker

debug logging on incremental workload

if counter is above reset counter federated workload

final compressRunFrequency

 decompressing mult print

colgroup uncompressed does not invalidate right multiplication

edge cases

aclculate sum to forcefully materialize count

add safety decompress on RMM.

fix conteres federated workload analysis

ColGroup Dense Selection

dense

+++ src/main/java/org/apache/sysds/runtime/instructions/spark/MultiReturnParameterizedBuiltinSPInstruction.java
@@ -317,7 +317,7 @@ public class MultiReturnParameterizedBuiltinSPInstruction extends ComputationSPI
 			Iterator<Object> iter = arg0._2().iterator();
 			ArrayList<String> ret = new ArrayList<>();
 
-			long rowID = 1;
+			int rowID = 1;
 			StringBuilder sb = new StringBuilder();
 
 			// handle recode maps
@@ -362,7 +362,7 @@ public class MultiReturnParameterizedBuiltinSPInstruction extends ComputationSPI
 			else {
 				throw new DMLRuntimeException("Unsupported metadata output for encoder: \n" + _encoder);
 			}
-			_accMax.add(rowID - 1);
+			_accMax.add(rowID - 1L);
 
 			return ret.iterator();
 		}
